+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - New call created returns the openvpn nodes with additional information such as the address.
+- Backup sql is compressed (zip) in order to save space.
+- Changed worker iteration in order to avoid iteration overlaps
 
 ### Fixed
 - Backup related tasks are not scheduled twice.